[#6] Move usecase to domain layer
