+++ buildKonfig.properties.example
