+++ shared/src/commonMain/kotlin/co/nimblehq/avishek/phong/kmmic/di/module/UseCaseModule.kt
@@ -1,6 +1,5 @@
 package co.nimblehq.avishek.phong.kmmic.di.module
 
-import co.nimblehq.avishek.phong.kmmic.data.usecase.*
 import co.nimblehq.avishek.phong.kmmic.domain.usecase.*
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
