+++ shared/src/commonMain/kotlin/co/nimblehq/avishek/phong/kmmic/domain/usecase/CheckLoggedInUseCase.kt
@@ -6,3 +6,12 @@ interface CheckLoggedInUseCase {
 
     operator fun invoke(): Boolean
 }
+
+class CheckLoggedInUseCaseImpl(
+    private val authenticationRepository: AuthenticationRepository
+): CheckLoggedInUseCase {
+
+    override fun invoke(): Boolean {
+        return authenticationRepository.hasCachedToken()
+    }
+}
